requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
codegen==1.0
coreapi==2.3.3
coreschema==0.0.4
datetime-truncate==1.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.13
django-admin-interface==0.19.0
django-admin-reset==0.5.2
django-admin-tools==0.9.2
django-appconf==1.0.5
django-ckeditor==6.4.0
django-colorfield==0.6.3
django-debug-toolbar==3.2.4
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-imagekit==4.1.0
django-import-export==2.8.0
django-jet==1.0.8
django-js-asset==2.0.0
django-jsonfield==1.4.1
django-multiselectfield==0.1.12
django-otp==1.1.3
django-qsstats-magic==1.1.0
django-recaptcha==3.0.0
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djcacheutils==3.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
feedparser==6.0.8
frozenlist==1.3.0
google-api-python-client==1.4.1
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupPy==1.14
MarkupSafe==2.1.1
model-bakery==1.5.0
multidict==6.0.2
oauth2client==4.1.3
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
patterns==0.3
pilkit==2.0
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
redis==4.2.2
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.16.0
slackclient==2.9.4
sqlparse==0.4.2
tablib==3.2.1
turtle==0.0.1
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
